# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Garuka API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/classes/{class_id}/attendance": {
      "get": {
        "tags": [
          "attendance"
        ],
        "summary": "Get attendance and student list for class on date",
        "operationId": "get_class_attendance",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Class Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassAttendanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "attendance"
        ],
        "summary": "Submit or update attendance for class on date",
        "operationId": "submit_class_attendance",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Class Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAttendanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceSubmissionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/absences/{absence_id}": {
      "delete": {
        "tags": [
          "attendance"
        ],
        "summary": "Void an absence record",
        "operationId": "void_absence",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "absence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Absence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attendance/compliance": {
      "get": {
        "tags": [
          "attendance"
        ],
        "summary": "Get class compliance grid for school and date range",
        "operationId": "get_attendance_compliance",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "school_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "School Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchoolComplianceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbsenceStatusEnum": {
        "type": "string",
        "enum": [
          "active",
          "voided"
        ],
        "title": "AbsenceStatusEnum"
      },
...
      "AttendanceSubmissionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "class_id": {
            "type": "string",
            "format": "uuid",
            "title": "Class Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "submitted_by": {
            "type": "string",
            "format": "uuid",
            "title": "Submitted By"
          },
          "source": {
            "$ref": "#/components/schemas/SubmissionSourceEnum"
          },
          "absent_count": {
            "type": "integer",
            "title": "Absent Count"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "class_id",
          "date",
          "submitted_by",
          "source",
          "absent_count",
          "submitted_at"
        ],
        "title": "AttendanceSubmissionResponse"
      },
...
      "ClassAttendanceResponse": {
        "properties": {
          "class_id": {
            "type": "string",
            "format": "uuid",
            "title": "Class Id"
          },
          "class_name": {
            "type": "string",
            "title": "Class Name"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "submission": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AttendanceSubmissionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "students": {
            "items": {
              "$ref": "#/components/schemas/StudentAttendanceItem"
            },
            "type": "array",
            "title": "Students",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "class_id",
          "class_name",
          "date"
        ],
        "title": "ClassAttendanceResponse"
      },
      "ClassCompliance": {
        "properties": {
          "class_id": {
            "type": "string",
            "format": "uuid",
            "title": "Class Id"
          },
          "class_name": {
            "type": "string",
            "title": "Class Name"
          },
          "grade": {
            "type": "integer",
            "title": "Grade"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/DayCompliance"
            },
            "type": "array",
            "title": "Days"
          }
        },
        "type": "object",
        "required": [
          "class_id",
          "class_name",
          "grade",
          "days"
        ],
        "title": "ClassCompliance"
      },
...
      "DayCompliance": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "absent_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Absent Count"
          }
        },
        "type": "object",
        "required": [
          "date",
          "status"
        ],
        "title": "DayCompliance"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReasonCodeEnum": {
        "type": "string",
        "enum": [
          "SICK",
          "WORK",
          "COST",
          "DISTANCE",
          "OTHER"
        ],
        "title": "ReasonCodeEnum"
      },
...
      "SchoolComplianceResponse": {
        "properties": {
          "school_id": {
            "type": "string",
            "format": "uuid",
            "title": "School Id"
          },
          "school_name": {
            "type": "string",
            "title": "School Name"
          },
          "from_date": {
            "type": "string",
            "format": "date",
            "title": "From Date"
          },
          "to_date": {
            "type": "string",
            "format": "date",
            "title": "To Date"
          },
          "compliance_pct": {
            "type": "number",
            "title": "Compliance Pct"
          },
          "classes": {
            "items": {
              "$ref": "#/components/schemas/ClassCompliance"
            },
            "type": "array",
            "title": "Classes"
          }
        },
        "type": "object",
        "required": [
          "school_id",
          "school_name",
          "from_date",
          "to_date",
          "compliance_pct",
          "classes"
        ],
        "title": "SchoolComplianceResponse"
      },
...
      "StudentAttendanceItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "roll_number": {
            "type": "integer",
            "title": "Roll Number"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "is_absent": {
            "type": "boolean",
            "title": "Is Absent"
          },
          "absence_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Absence Id"
          },
          "reason_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReasonCodeEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AbsenceStatusEnum"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "roll_number",
          "full_name",
          "is_absent"
        ],
        "title": "StudentAttendanceItem"
      },
...
      "SubmissionSourceEnum": {
        "type": "string",
        "enum": [
          "ussd",
          "dashboard"
        ],
        "title": "SubmissionSourceEnum"
      },
      "SubmitAttendanceRequest": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "absent_student_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Absent Student Ids",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "SubmitAttendanceRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}